requirements.txt
# YOLOv5-PyTorch Requirements
# Python >= 3.8

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Data Processing
numpy>=1.20.0
PyYAML>=5.4.0

# Training & Evaluation
tqdm>=4.60.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
